Set C binding for unix_lockf_job in its own C file
