+++ src/unix/jbuild
@@ -193,6 +193,7 @@
     unix_readdir_job
     unix_readdir_n_job
     unix_readlink_job
+    unix_lockf_job
     not_available
   ))
   (install_c_headers (
